requirements.txt
mcp>=0.1.0
mcp-use>=0.1.0
langchain-openai>=0.0.2
python-dotenv>=1.0.0
openai>=1.0.0 